删除无用种子文件

+++ src/helper/TorrentHelper.js
@@ -1,6 +1,6 @@
 let MongoUtil = require('../utils/MongoUtil');
 let Torrent = require('../model/Torrent');
-let TorrentParseUtil = require('../utils/TorrentParseUtil');
+let parseTorrentFile = require('../utils/TorrentParseUtil');
 let fs = require("fs");
 let logger = require('pomelo-logger').getLogger('TorrentHelper.js');
 
@@ -12,23 +12,33 @@ const getPath = (hash, hasSuffix) => {
     return `${getParentPath()}${hash}${hasSuffix ? '' : '.torrent'}`
 };
 
-const saveTorrent = (hash, hasSuffix) => {
-    let torrent = TorrentParseUtil(getPath(hash, hasSuffix));
+const getTorrent = (hash, hasSuffix) => {
+    let torrent = parseTorrentFile(getPath(hash, hasSuffix));
     torrent._id = torrent.infoHash;
-    return MongoUtil.insertDocUnique(Torrent, torrent)
+    return torrent;
 };
 
 const saveTorrentList = async (torrentList) => {
     let fileName = torrentList.pop();
     if (!!fileName && fileName.indexOf('.torrent') > 0) {
         try {
-            let result = await saveTorrent(fileName, true);
-            if (result && result.msg) {
-                logger.info(result.msg);
-            } else {
-                logger.info(`${fileName} => ${!!result ? '保存成功' : '保存失败'}`);
+            let torrent;
+            try {
+                torrent = getTorrent(fileName, true);
+            } catch (e) {
+                logger.error(e);
+                logger.error(fileName);
+                fs.unlinkSync(getParentPath() + fileName);
             }
-            fs.unlinkSync(getParentPath() + fileName);
+            if (!!torrent) {
+                let result = await MongoUtil.insertDocUnique(Torrent, torrent);
+                if (result && result.msg) {
+                    logger.info(result.msg);
+                } else {
+                    logger.info(`${fileName} => ${!!result ? '保存成功' : '保存失败'}`);
+                }
+                fs.unlinkSync(getParentPath() + fileName);
+            }
             await saveTorrentList(torrentList);
         } catch (e) {
             logger.error(e);
@@ -48,7 +58,7 @@ const scanAndSaveTorrent = () => {
         let page = Math.ceil(length / 500);
         for (let i = 0; i < page; i++) {
             await saveTorrentList(files.slice(i * 500, (i + 1) * 500));
-            console.log(`${i / page}%`);
+            console.log(`${(i + 1) / page}%`);
         }
         MongoUtil.close();
     });
@@ -57,7 +67,8 @@ const scanAndSaveTorrent = () => {
 MongoUtil.connect();
 scanAndSaveTorrent();
 
+
 module.exports = {
-    saveTorrent,
+    saveTorrent: getTorrent,
     scanAndSaveTorrent
 };
